build: Safe removal of the build directory
Never rm -rf a directory without checking anything. The tool will
happily rm anything you throw at it, including
"${EMPTYVAR1}/${EMPTYVAR2}" resolving to rm -rf /.

Note that in this specific case, shellcheck will warn you and suggest
using "${EMPTYVAR1:?}" to avoid var1 from being empty. Obviously this
won't be trigged when doing PATH="${EMPTYVAR1}/${EMPTYVAR2}"; rm -rf
"${PATH}".

By using a fixed name template and check for its existence we can
_reduce_ chances of breakage.

Contributes to EMP-416

+++ build.sh
@@ -17,7 +17,8 @@ ALPINE_REPO="${ALPINE_REPO:-http://dl-cdn.alpinelinux.org/alpine}"
 TOOLBOX_IMAGE="${TOOLBOX_IMAGE:-um-update_toolbox.xz.img}"
 
 CUR_DIR="$(pwd)"
-BUILD_DIR="${CUR_DIR}/.build_${ARCH}"
+NAME_TEMPLATE_BUILD_DIR=".build_${ARCH}"
+BUILD_DIR="${BUILD_DIR:-${CUR_DIR}/${NAME_TEMPLATE_BUILD_DIR}}"
 ROOTFS_DIR="${BUILD_DIR}/rootfs"
 
 SYSTEM_UPDATE_DIR="/etc/system_update"
@@ -45,7 +46,9 @@ cleanup()
         exit 1
     fi
 
-    rm -rf "${BUILD_DIR:?}"
+    if [ -z "${BUILD_DIR##*${NAME_TEMPLATE_BUILD_DIR}*}" ]; then
+        rm -rf "${BUILD_DIR:?}"
+    fi
 }
 
 bootstrap_prepare()
